Allow openjade to be built against the standalone, independent opensp (currently 1.5).
Note two extract options to configure

--spincludedir=<path to OpenSP include directory>
--splibdir=<path to OpenSP library directory>

These default to /usr/include/OpenSP and /usr/lib because that is
where they are on my system.

+++ jade/grove/Node.h
@@ -37,7 +37,7 @@ namespace GROVE_NAMESPACE {
 #ifdef SP_WCHAR_T_USHORT
 typedef wchar_t GroveChar;
 #else
-typedef unsigned short GroveChar;
+typedef unsigned int GroveChar;
 #endif
 #else /* not SP_MULTI_BYTE */
 typedef unsigned char GroveChar;
